+++ packages/web3/src/crypto-input/constants.ts
@@ -1,0 +1,39 @@
+/**
+ * CryptoInput overwrite InputFontSizeSM token
+ */
+const INPUT_FONT_SIZE_SM = 20;
+
+/**
+ * CryptoInput overwrite InputFontSize token
+ */
+const INPUT_FONT_SIZE = 24;
+
+/**
+ * CryptoInput overwrite InputFontSizeLG token
+ */
+const INPUT_FONT_SIZE_LG = 28;
+
+/**
+ * CryptoInput overwrite PaddingInline token
+ */
+const PADDING_INLINE = 0;
+
+/**
+ * CryptoInput overwrite PaddingBlock token
+ */
+const PADDING_BLOCK = 0;
+
+/**
+ * CrpytoInput default Token
+ */
+export const CRYPTO_INPUT_TOKEN = {
+  inputFontSizeSM: INPUT_FONT_SIZE_SM,
+  inputFontSize: INPUT_FONT_SIZE,
+  inputFontSizeLG: INPUT_FONT_SIZE_LG,
+  paddingInlineSM: PADDING_INLINE,
+  paddingInline: PADDING_INLINE,
+  paddingInlineLG: PADDING_INLINE,
+  paddingBlockSM: PADDING_BLOCK,
+  paddingBlock: PADDING_BLOCK,
+  paddingBlockLG: PADDING_BLOCK,
+};