+++ packages/web3/src/crypto-input/demos/customHeader.tsx
@@ -1,0 +1,20 @@
+import React, { useState } from 'react';
+import { CryptoInput, type CryptoInputProps } from '@ant-design/web3';
+import { ETH, USDT } from '@ant-design/web3-assets/tokens';
+
+const App: React.FC = () => {
+  const [crypto, setCrypto] = useState<CryptoInputProps['value']>();
+
+  return (
+    <div style={{ width: 456 }}>
+      <CryptoInput
+        value={crypto}
+        onChange={setCrypto}
+        header={'Crypto Input Header'}
+        tokenList={[ETH, USDT]}
+      />
+    </div>
+  );
+};
+
+export default App;